MODFISTO-345: Remove update_encumbrances_order_status.ftl (cross module)
Remove update_encumbrances_order_status.ftl, a migration script for Iris (R1 2021) that makes illegal cross-module database schema reads.

Problem:

The migration script

https://github.com/folio-org/mod-finance-storage/blob/v7.0.6/src/main/resources/templates/db_scripts/migration/update_encumbrances_order_status.ftl

reads from ${myuniversity}_mod_orders_storage. This is an illegal cross-module database schema read. It is also executed when enabling mod-finance-storage for a tenant - at that time mod-orders-storage might not have been enabled for the tenant resulting in a failure.

This migration script is used to migrate from a version before Iris to Iris (R1 2021) or any later version. However, Iris (R1 2021) reached end of life and is out of support since the release of Kiwi (R3 2021) on 2022-01-13.

Therefore the migration is no longer needed and should be deleted. This fixes the illegal cross-module schema read problem.